requirements.txt
alembic==1.10.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
greenlet==2.0.2
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mysqlclient==2.1.1
oauthlib==3.2.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
